RepoBiteCoFinal/biteco_fixed/Seguridad/requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
django-redis
redis
PyJWT
